+++ tests/base.py
@@ -85,6 +85,14 @@ class HubspotBaseTest(BaseCase):
                 self.EXPECTED_PAGE_SIZE: 250,
                 self.OBEYS_START_DATE: True
             },
+            "list_memberships": {
+                self.PRIMARY_KEYS: {"recordId", "listId"},
+                self.REPLICATION_METHOD: self.INCREMENTAL,
+                self.REPLICATION_KEYS: {"membershipTimestamp"},
+                self.EXPECTED_PAGE_SIZE: 250,
+                self.OBEYS_START_DATE: True,
+                self.PARENT_STREAM: 'contact_lists'
+            },
             "contacts": {
                 self.PRIMARY_KEYS: {"id"},
                 self.REPLICATION_METHOD: self.INCREMENTAL,
@@ -130,6 +138,13 @@ class HubspotBaseTest(BaseCase):
                 self.REPLICATION_KEYS: {"updatedAt"},
                 self.OBEYS_START_DATE: True
             },
+            "form_submissions": {
+                self.PRIMARY_KEYS: {"conversionId"},
+                self.REPLICATION_METHOD: self.INCREMENTAL,
+                self.REPLICATION_KEYS: {"submittedAt"},
+                self.OBEYS_START_DATE: True,
+                self.PARENT_STREAM: 'forms'
+            },
             "owners": {
                 self.PRIMARY_KEYS: {"id"},
                 self.REPLICATION_METHOD: self.INCREMENTAL,
@@ -156,35 +171,6 @@ class HubspotBaseTest(BaseCase):
                 self.EXPECTED_PAGE_SIZE: 100,
                 self.OBEYS_START_DATE: True
             },
-            # below are the custom_objects stream
-            "cars": {
-                self.PRIMARY_KEYS: {"id"},
-                self.REPLICATION_METHOD: self.INCREMENTAL,
-                self.REPLICATION_KEYS: {"updatedAt"},
-                self.EXPECTED_PAGE_SIZE: 100,
-                self.OBEYS_START_DATE: True
-            },
-            "co_firsts": {
-                self.PRIMARY_KEYS: {"id"},
-                self.REPLICATION_METHOD: self.INCREMENTAL,
-                self.REPLICATION_KEYS: {"updatedAt"},
-                self.EXPECTED_PAGE_SIZE: 100,
-                self.OBEYS_START_DATE: True
-            },
-            "custom_object_campaigns": {
-                self.PRIMARY_KEYS: {"id"},
-                self.REPLICATION_METHOD: self.INCREMENTAL,
-                self.REPLICATION_KEYS: {"updatedAt"},
-                self.EXPECTED_PAGE_SIZE: 100,
-                self.OBEYS_START_DATE: True
-            },
-            "custom_object_contacts": {
-                self.PRIMARY_KEYS: {"id"},
-                self.REPLICATION_METHOD: self.INCREMENTAL,
-                self.REPLICATION_KEYS: {"updatedAt"},
-                self.EXPECTED_PAGE_SIZE: 100,
-                self.OBEYS_START_DATE: True
-            }
         }
 
     #############################
@@ -224,7 +210,10 @@ class HubspotBaseTest(BaseCase):
     
     def failed_sync_streams(self):
         """A set of streams that are known to have sync issues and should be skipped in assertion errors."""
-        return {"subscription_changes", "email_events"}
+        return {"email_events",
+                "form_submissions",  # Added Apr 2026, client creates not yet implemented
+                "subscription_changes",
+                }
     
     def validate_failed_sync_streams(self, stream, *synced_data_records):
         """